pyproject.toml
[tool.poetry]
name = "fastapiauth"
version = "0.1.0"
description = ""
authors = ["DanyloSolohub <danylo.solohub.te.2019@lpnu.ua>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.85.1"
uvicorn = "^0.19.0"
python-dotenv = "^0.21.0"
SQLAlchemy = "^1.4.42"
asyncpg = "^0.26.0"
alembic = "^1.8.1"
SQLAlchemy-Utils = "^0.38.3"
psycopg2 = "^2.9.4"
pydantic = {extras = ["email"], version = "^1.10.2"}
fastapi-pagination = "^0.10.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
